revno: 15527 message:   Fixed bug - Exception thrown when to delete a program
committer: Tran Chau <tran.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Tue 2014-06-03 12:55:04 +0800

+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/program/ProgramInstanceDeletionHandler.java
@@ -28,6 +28,7 @@ package org.hisp.dhis.program;
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import java.util.Collection;
 import java.util.Iterator;
 
 import org.hisp.dhis.system.deletion.DeletionHandler;
@@ -117,13 +118,17 @@ public class ProgramInstanceDeletionHandler
     @Override
     public void deleteProgram( Program program )
     {
-        Iterator<ProgramInstance> iterator = program.getProgramInstances().iterator();
+        Collection<ProgramInstance> programInstances = programInstanceService.getProgramInstances( program );
 
-        while ( iterator.hasNext() )
+        if ( programInstances != null )
         {
-            ProgramInstance programInstance = iterator.next();
-            iterator.remove();
-            programInstanceService.deleteProgramInstance( programInstance );
+            Iterator<ProgramInstance> iterator = programInstances.iterator();
+            while ( iterator.hasNext() )
+            {
+                ProgramInstance programInstance = iterator.next();
+                iterator.remove();
+                programInstanceService.deleteProgramInstance( programInstance );
+            }
         }
     }
 }
